Remove unreliable file written in C: mpu6050 driver , quaternion math Remove test task pid, read mpu6050, printf

+++ Src/quaternion_math.cpp
@@ -1,5 +1,5 @@
 #include "quaternion.hpp"
-#include "myMath.hpp"
+#include "quat_math.hpp"
 
 #include <math.h>
 
@@ -35,8 +35,8 @@ float quat_get_Yaw(const quaternion& q) {
 quaternion quat2EulerAngle(const quaternion& q) {
     return quaternion{
         0,
-        atan((2*q[0]*q[1] + 2*q[2]*q[3]) / (1-2*q[1]*q[1] - 2*q[2]*q[2])),
-        -asin(2*q[1]*q[3]-2*q[0]*q[2]),
-        atan((2*q[1]*q[2] + 2*q[0]*q[3]) / (1-2*q[2]*q[2] - 2*q[3]*q[3]))
+        (float)atan((2*q[0]*q[1] + 2*q[2]*q[3]) / (1-2*q[1]*q[1] - 2*q[2]*q[2])),
+        (float)-asin(2*q[1]*q[3]-2*q[0]*q[2]),
+        (float)atan((2*q[1]*q[2] + 2*q[0]*q[3]) / (1-2*q[2]*q[2] - 2*q[3]*q[3]))
     };
 }
